+++ changelog.md
@@ -6,7 +6,7 @@
 * `.econ` now comments out blocks if no `BaseType`s are found for the rule parameters
 * Added the new `.import` handler, which allows importing of entire filters, blocks, or lines within blocks
 * Replaced `.chaos` with `.tag`, a much more generic version of the same idea
-* Added `Filter` and `Line` as new abstractions to provide more flexibility and code reuse to handlers.
+* Added `Filter` and `Line` as new abstractions to provide more flexibility and code reuse to handlers
 * Renamed sections to blocks as that's the GGG standard name
 * Refactored and cleaned up the code significantly, so that future work is simpler
 * Used single underscore for private instead of double underscore